\name{smutans.feature.genes}
\docType{data}
\alias{smutans.feature.genes}
\title{Feature table of genes}
\description{
\code{smutans.feature.genes} is a table of six columns of chrmosome, start, end,
gene name, 0, and strand.
}
\usage{
data("smutans.feature.genes")
}

\format{ExonCountSet, CountDataSet}

\source{Processed data from NCBI Gene Expression Omnibus under accession numbers GSM461176 to GSM461181.}
\references{Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010}
\keyword{datasets}
