\name{smutans.go.genes}
\docType{data}
\alias{smutans.go.genes}
\title{Association table of gene ontology terms and genes}
\description{
\code{smutans.go.genes} is a table of three columns of gene name, functional
category, and a p-value: e.g., SMU.01 GO:0000166 0.
}
\usage{
data("smutans.go.genes")
}

\format{ExonCountSet, CountDataSet}

\source{Processed data from NCBI Gene Expression Omnibus under accession numbers GSM461176 to GSM461181.}
\references{Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010}
\keyword{datasets}
