\name{smutans.keg.desc}
\docType{data}
\alias{smutans.keg.desc}
\title{Functional category table}
\description{
\code{smutans.keg.desc} is a table of three columns of functional category,
number of genes, and category description: e.g., GO:0007205 1 alcohol metabolic
process.
}
\usage{
data("smutans.keg.desc")
}

\format{ExonCountSet, CountDataSet}

\source{Processed data from NCBI Gene Expression Omnibus under accession numbers GSM461176 to GSM461181.}
\references{Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010}
\keyword{datasets}
