\name{smutansGenes}
\docType{data}
\alias{smutansGenes}
\title{Read counts or per gene from RNA-seq samples}
\description{
\code{smutansGenes} is an \code{ExonCountSet} object containing gene counts for each of the samples from RNA-seq data.
\code{smutansGenes} is an \code{CountDataSet} object with gene level counts.
}
\usage{
data("smutansGenes")
}

\format{ExonCountSet, CountDataSet}

\source{Processed data from NCBI Gene Expression Omnibus under accession numbers GSM461176 to GSM461181.}
\references{Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010}
\keyword{datasets}
