\name{smutans.genes.criteria}
\docType{data}
\alias{smutans.genes.criteria}
\title{Functional category table}
\description{
\code{smutans.genes.criteria} is a table of values for targets of putative small
RNAs.}
\usage{
data("smutans.genes.criteria")
}
\source{Processed data using RNAz, RNAplex, and RNAplfold.}
\references{Choi et al. Defining Transcriptome using RNA Sequencing of Streptococcus mutans. Genome Research, 2012}
\keyword{datasets}
