\name{smutans.title}
\Rdversion{1.1}
\alias{smutans.title}

\title{
   Name of the Smutan object.
}
\description{
   This function has access to the title of an Smutans object.
}
\usage{
smutans.title(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
      a Smutans with an experimental design
}
}
\details{
   See \code{\link{smutans.title}} for more technical informations
}
\value{
A string.
}
\author{
   Sang Chul Choi, schoi@cornell.edu
}
\examples{
%data( "smutansGenes" )
%x <- newSmutans(smutansGenes)
%x <- smutans.title(x)
}
